backend/requirements.txt
Flask==2.2.5
requests==2.31.0
Werkzeug==2.2.3
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7

waitress==2.1.2

psycopg2-binary==2.9.9

numpy==1.21.6
scipy==1.7.3
scikit-learn==1.0.2
joblib==1.3.2
threadpoolctl==3.1.0

matplotlib==3.5.3
cycler==0.11.0
fonttools==4.38.0
kiwisolver==1.4.4
packaging==23.2
Pillow==9.5.0
pyparsing==3.1.4
python-dateutil==2.8.2
six==1.16.0